Refactored Fetch back to AssetPrice
